display_scatter move 2d and 3d logic to subclasses
